requirements.txt
flask
gunicorn==20.1.0 
python-dotenv
